Add files via upload
Atoti library in python is used to provide the user interface to view data as per their wish and the analyzed data using pandas, NumPy, and other libraries facilitated in better understanding of the plots when plotted against a different model of testing. The labeling of the data is taken care of by atoti itself and separate labeling isn't necessary. The Pycaret was used for the auto-ml part.